[lection-3] - правки по комментам 2
