+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/dbUser/DBUser.java
@@ -9,7 +9,7 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-@ExtendWith({DBUserExtension.class, DBUserExtensionParameterResolver.class})
+@ExtendWith(DBUserExtension.class)
 public @interface DBUser {
 
 	String username();
