+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/DBUserTest.java
@@ -1,5 +1,7 @@
 package guru.qa.niffler.test;
 
+import com.codeborne.selenide.Selenide;
+import guru.qa.niffler.db.model.UserEntity;
 import guru.qa.niffler.jupiter.dao.DaoExtension;
 import guru.qa.niffler.jupiter.dbUser.DBUser;
 import org.junit.jupiter.api.Test;
@@ -11,12 +13,17 @@ import static com.codeborne.selenide.Selenide.$;
 @ExtendWith(DaoExtension.class)
 public class DBUserTest extends BaseWebTest {
 
-	private static final String USERNAME = "Adam Johnson";
+	private static final String USERNAME = "Adam 5";
 	private static final String PASSWORD = "12345678";
 
 	@DBUser(username = USERNAME, password = PASSWORD)
-	@Test
-	void mainPageShouldBeVisibleAfterLogIn() {
+	@Test()
+	void mainPageShouldBeVisibleAfterLogIn(UserEntity user) {
+		Selenide.open("http://127.0.0.1:3000/main");
+		$("a[href*='redirect']").click();
+		$("input[name='username']").setValue(user.getUsername());
+		$("input[name='password']").setValue(user.getPassword());
+		$("button[type='submit']").click();
 		$(".main-content .main-content__section-stats").shouldBe(visible);
 	}
 }
